# Back/src/main/java/com/company/favorite/news/dto/ErrorResponseDTO.java
package com.company.favorite.news.dto;


import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class ErrorResponseDTO {
    String message;
}